LPS-74897 Rename package
