+++ modules/apps/foundation/petra/petra-lang/src/main/java/com/liferay/petra/lang/AutoResetThreadLocal.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.util;
+package com.liferay.petra.lang;
 
 import java.util.function.Supplier;
